requirements.txt
browser-use==0.2.2
fastapi==0.115.12
uvicorn==0.34.2
pytest==8.3.5
playwright==1.52.0
supabase==2.15.1
debugpy
psutil>=7.0.0
python-dotenv>=1.0.0
